split config classes
