+++ src/main/java/lt/liutikas/reddit/config/WebConfig.java
@@ -1,18 +1,10 @@
 package lt.liutikas.reddit.config;
 
-import com.azure.ai.textanalytics.TextAnalyticsClient;
-import com.azure.ai.textanalytics.TextAnalyticsClientBuilder;
-import com.azure.core.credential.AzureKeyCredential;
-import lt.liutikas.reddit.config.properties.AzureProperties;
-import lt.liutikas.reddit.config.properties.TwitterProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.MediaType;
 import org.springframework.web.servlet.config.annotation.ContentNegotiationConfigurer;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import twitter4j.Twitter;
-import twitter4j.TwitterFactory;
-import twitter4j.conf.ConfigurationBuilder;
 
 import java.time.Clock;
 
@@ -25,14 +17,6 @@ public class WebConfig {
     }
 
     @Bean
-    public TextAnalyticsClient textAnalyticsClient(AzureProperties properties) {
-        return new TextAnalyticsClientBuilder()
-                .credential(new AzureKeyCredential(properties.getKey()))
-                .endpoint(properties.getEndpoint())
-                .buildClient();
-    }
-
-    @Bean
     public WebMvcConfigurer myWebMvcConfigurer() {
         return new WebMvcConfigurer() {
             @Override
@@ -42,17 +26,5 @@ public class WebConfig {
         };
     }
 
-    @Bean
-    public Twitter twitter(TwitterProperties properties) {
-        ConfigurationBuilder cb = new ConfigurationBuilder();
 
-        cb.setDebugEnabled(true)
-                .setOAuthConsumerKey(properties.getConsumerKey())
-                .setOAuthConsumerSecret(properties.getConsumerSecret())
-                .setOAuthAccessToken(properties.getAccessToken())
-                .setOAuthAccessTokenSecret(properties.getAccessTokenSecret());
-        TwitterFactory tf = new TwitterFactory(cb.build());
-
-        return tf.getInstance();
-    }
 }
